services/profile/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
